Refactor popover logic
