+++ client-course-schedulizer/src/components/reuseables/AddNonTeachingLoadPopover/AddNonTeachingLoadPopover.scss
@@ -1,0 +1,11 @@
+.popover-container {
+  padding: 0.7em;
+}
+
+.popover-title {
+  font-weight: 600 !important;
+}
+
+.popover-buttons > * {
+  margin-right: 1em !important;
+}
