+++ client-course-schedulizer/src/components/reuseables/AddNonTeachingLoadPopover/AddNonTeachingLoadPopover.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const AddNonTeachingLoadPopover = () => {
+  return <></>;
+};
