+++ client-course-schedulizer/src/components/reuseables/AddNonTeachingLoadPopover/index.ts
@@ -1,0 +1,1 @@
+export * from "./AddNonTeachingLoadPopover";
