--- client-course-schedulizer/src/components/reuseables/AddSectionButton/AddSectionButton.scss
@@ -1,4 +1,0 @@
-.popover {
-  max-width: 50%;
-  min-width: 500px;
-}
