+++ client-course-schedulizer/src/components/reuseables/AddSectionButton/AddSectionButton.tsx
@@ -1,8 +1,7 @@
-import { Button, ButtonProps, IconButton, Popover } from "@material-ui/core";
-import { Add } from "@material-ui/icons";
+import { ButtonProps } from "@material-ui/core";
 import { AddSectionPopover } from "components";
-import { bindPopover, bindTrigger, usePopupState } from "material-ui-popup-state/hooks";
 import React from "react";
+import { PopoverButton } from "../PopoverButton";
 import "./AddSectionButton.scss";
 
 interface AddSectionButton extends ButtonProps {
@@ -10,38 +9,12 @@ interface AddSectionButton extends ButtonProps {
 }
 
 export const AddSectionButton = (props: AddSectionButton) => {
-  const popupState = usePopupState({
-    popupId: "addSection",
-    variant: "popover",
-  });
-  const { isIcon, ...buttonProps } = props;
+  const { isIcon } = props;
 
   return (
-    <>
-      {isIcon ? (
-        <IconButton {...bindTrigger(popupState)}>
-          <Add />
-        </IconButton>
-      ) : (
-        <Button color="secondary" variant="contained" {...buttonProps} {...bindTrigger(popupState)}>
-          Add Section
-        </Button>
-      )}
-      <Popover
-        {...bindPopover(popupState)}
-        anchorOrigin={{
-          horizontal: "left",
-          vertical: "bottom",
-        }}
-        PaperProps={{ style: { minWidth: "500px" } }}
-        transformOrigin={{
-          horizontal: "right",
-          vertical: "top",
-        }}
-      >
-        <AddSectionPopover />
-      </Popover>
-    </>
+    <PopoverButton isIcon={isIcon}>
+      <AddSectionPopover />
+    </PopoverButton>
   );
 };
