+++ client-course-schedulizer/src/components/reuseables/PopoverButton/PopoverButton.scss
@@ -1,0 +1,4 @@
+.popover {
+  max-width: 50%;
+  min-width: 500px;
+}
