+++ client-course-schedulizer/src/components/reuseables/PopoverButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./PopoverButton";
